Reorganised and abstracted certain function logic
As a result our application is slightly slower. One main cause is
`get_industry_overview` in `src/services/finance` due to not using
caching.
